+++ app/src/main/AndroidManifest.xml
@@ -32,7 +32,7 @@
             android:theme="@style/FinestWebViewTheme" />
 
         <service
-            android:name=".ListenService"
+            android:name=".presentation.features.listens.submitter.ListenService"
             android:enabled="true"
             android:exported="true"
             android:permission="android.permission.BIND_NOTIFICATION_LISTENER_SERVICE">
