+++ app/src/main/java/org/metabrainz/android/App.kt
@@ -8,11 +8,11 @@ import android.graphics.Typeface
 import android.net.ConnectivityManager
 import android.net.NetworkCapabilities
 import android.os.Build
-import android.os.Environment
 import android.provider.Settings
 import dagger.hilt.android.HiltAndroidApp
 import org.metabrainz.android.presentation.Configuration
 import org.metabrainz.android.presentation.UserPreferences.preferenceListeningEnabled
+import org.metabrainz.android.presentation.features.listens.submitter.ListenService
 
 @HiltAndroidApp
 class App : Application() {
