+++ app/src/main/java/org/metabrainz/android/presentation/features/listens/submitter/ListenSessionListener.kt
@@ -1,4 +1,4 @@
-package org.metabrainz.android
+package org.metabrainz.android.presentation.features.listens.submitter
 
 import android.media.MediaMetadata
 import android.media.session.MediaController
